Split out core gateway functions from D-Bus interface
